+++ dConnectDevicePlugin/dConnectDeviceLinking/app/src/main/java/org/deviceconnect/android/deviceplugin/linking/linking/profile/LinkingTemperatureProfile.java
@@ -70,7 +70,7 @@ public class LinkingTemperatureProfile extends TemperatureProfile {
         TemperatureType tType = TemperatureType.TYPE_CELSIUS;
         float temp = temperature;
         if (type == TemperatureType.TYPE_FAHRENHEIT.getValue()) {
-            temp = TemperatureProfile.convertCelsiusToFahrenheit((int) temp);
+            temp = TemperatureProfile.convertCelsiusToFahrenheit(temp);
             tType = TemperatureType.TYPE_FAHRENHEIT;
         }
         setResult(response, DConnectMessage.RESULT_OK);
